backend/requirements.txt
Django==5.1.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.1.1
image==1.5.33
itsdangerous==2.2.0
jinja2==3.1.4
numpy==2.1.3
pillow==11.0.0
pip==24.3.1
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wtforms==3.2.1
requests==2.31.0
pyodm
opencv-python
opencv-contrib-python